frontend/requirements.txt
streamlit
python-dotenv
solana
requests